refactor create option modal
