+++ Composer/packages/client/src/components/CreationFlow/v2/CreationFlow.tsx
@@ -98,6 +98,11 @@ const CreationFlowV2: React.FC<CreationFlowProps> = () => {
     navigate(`/home`);
   };
 
+  const handleJumpToOpenModal = () => {
+    setCreationFlowStatus(CreationFlowStatus.OPEN);
+    navigate('./open');
+  };
+
   const openBot = async (botFolder) => {
     setCreationFlowStatus(CreationFlowStatus.CLOSE);
     await openProject(botFolder, 'default', true, (projectId) => {
@@ -175,6 +180,7 @@ const CreationFlowV2: React.FC<CreationFlowProps> = () => {
           path="create"
           templates={templateProjects}
           onDismiss={handleDismiss}
+          onJumpToOpenModal={handleJumpToOpenModal}
           onNext={handleCreateNext}
         />
         <DefineConversationV2
